Move 2 functions to uncertainty utils
